Made test for cicle area and circle perimeter
Changed Point2D to use double as coordinates instead of int values. Added Point2DTest to test distance method

+++ src/Shape.java
@@ -2,7 +2,7 @@ public abstract class Shape {
     ShapeType type;
     Point2D center;
 
-    Shape(int x, int y,ShapeType shapeType){
+    Shape(double x, double y,ShapeType shapeType){
         this.type=shapeType;
         center=new Point2D(x,y);
     }
